Improve wx gui backend
