+++ examples/triangle_qt.py
@@ -3,6 +3,7 @@ Import the viz from triangle.py and run it in a Qt window.
 Works with either PyQt6, PyQt5 or PySide2.
 """
 
+# For the sake of making this example Just Work, we try multiple QT libs
 try:
     from PyQt6 import QtWidgets
 except ModuleNotFoundError:
@@ -10,21 +11,20 @@ except ModuleNotFoundError:
         from PyQt5 import QtWidgets
     except ModuleNotFoundError:
         from Pyside2 import QtWidgets
+
 from wgpu.gui.qt import WgpuCanvas  # WgpuCanvas is a QWidget subclass
 import wgpu.backends.rs  # noqa: F401, Select Rust backend
 
-# Import the (async) function that we must call to run the visualization
-from triangle import main
+from triangle import main  # The function to call to run the visualization
 
 
 app = QtWidgets.QApplication([])
 canvas = WgpuCanvas(title="wgpu triangle with Qt")
 
 main(canvas)
-try:
-    app.exec()
-except Exception:
-    app.exec_()
+
+# Enter Qt event loop (compatible with qt5/qt6)
+app.exec() if hasattr(app, "exec") else app.exec_()
 
 
 # For those interested, this is a simple way to integrate Qt's event
